refactor to combine network fee selector with gasless switch
